refactor: remove JPG5 crawler
